Updated openstack/openstack Project: openstack/nova-specs  861c4c634660d21c4c577251ed4fa4616f3d9c12
Amend placement REST API with allocation records

Not sure how or why we forgot to include an ability for a caller to set
and delete allocation records against resource providers, but we did.

This amendment to the spec adds two additional REST API calls for
setting and deleting allocation records in an atomic fashion (atomic
operation for a set of resources consumed by a single consumer against
one or more resource providers). These REST API calls will be made from
the Nova resource tracker when instances claim or free resources on the
compute nodes.

The endpoint is /allocations/{consumer_uuid} accepting PUT and DELETE
methods.

Co-Authored-By: Chris Dent <cdent@anticdent.org>
Change-Id: I5ef70ae2e13b1dd0c0ef1fa5821b3ee5a3265ca7
Blueprint: generic-resource-pools