Updated openstack/openstack Project: openstack/nova  0090e660d54e69d25f2a76222553a0661b0b1699
instance.name should be blank if instance.id is not set

By default the ability to generate instance.name depends on the
instance.id field. However the value for this field comes from the
autoincrement id column in the instance table. This means that if the
instance has not been written to a database, as will be the case before
scheduling in cellsv2, there is no way to properly determine what the
instance name will be.

In that situation the instance name will now be returned as '' if the
vm_state and task_state match the pre-scheduled instance case. If the
instance does not have the right vm_state/task_state set an error will
be logged and the instance uuid will be returned as the name. Using uuid
matches what happens for template failures.

Change-Id: I4a743c61a619af6f91a102a219fc0ec49df19d9d