Updated openstack/openstack Project: openstack/tripleo-heat-templates  dd8cb29d4e598a51fee04a4b6d8cf4a91cb82d97
Support arbitrary attributes for the Ceph pools

This change adds a CephPools parameter which can be used to
provide custom settings for any Ceph pool.

It also removes our custom setting for the default pg_num,
pgp_num and size so that these are used for the managed
pools (and can be overridden) but aren't enforced globally.

Change-Id: Idcf28bec46beabb1b590fc8e78b43e58d8e35717
Closes-Bug: 1517969
Depends-On: I38978f0f3119e4ab7dd45021e598253cb066cb5a