Updated openstack/openstack Project: openstack/horizon  5dbbc329ca576dc1f121a4e9ff7ab9f5a146e553
hz-resource-table handles transitional states

In the django tables, if a resource goes into a transitional-state
(think building, deleting, shutting-down, ...) the table would highlight
that behavior with 2 changes:
A) the row would be highlighted
B) a barber-pole would show up in the status column

The new-ish registry-based hz-resource-table wasn't making either of
those visual changes.  But now it is!  Hurray for fewer regressions!

Change-Id: Ic84b1eabc2c32e7ba48bf9ce2279f69a4e9a95e0
Partially-Implements: blueprint angular-registry
Co-Authored-By: Tyr Johanson tyr@hpe.com