Updated openstack/openstack Project: openstack/neutron  fd401fe0a052a7103cb19d7385a1c702de05577f
Rollback router intf port update if csnat update fails

If a CSNAT port update fails when a subnet is added,
it will leave the regular updated router port in the
changed state since it's a separate port update with
a separate transaction. So when the CSNAT update fails
we need explicitly undo the 1st update with another update.

Closes-Bug: #1609693
Change-Id: If979aebbebf6da77f3fcec0cc7ec8b27e388aba4