Updated openstack/openstack Project: openstack-infra/project-config  9337c700adc04145b9dbb5e00805c17e2387339a
Only error for duplicate requirements when strict

If a project adds a requirements check job while its requirements
files are nonconforming, they should be able to fix it by submitting
a change which fixes their requirements. If validity checks get
performed while reading requirements lists in the
gate{name}-requirements job, they need to only trigger failure while
parsing the submitted requirements and not on the existing state of
requirements before applying a proposed change.

Pass the "strict" flag through and wrap the duplicate entries check
in a conditional for it, so that we don't block fixes for projects
that start out with a duplicate requirement.

Change-Id: I16cbea930fd104717e73bd4e012ec77fc4f98494