Updated openstack/openstack Project: openstack/cinder  1eaa23098cd74f1c5f2ecf1814af7ad2350b2add
Remove resource_common_manage load warning

The file resource_common_manage.py contains helper code, but the
extension loader is trying to load it. Loading fails because it is not
an extension and a warning is printed to the log. This patch causes this
file to be skipped to avoid adding noise to the logs.

Change-Id: I138793869d8372cd292b28ae44eaa193f39a8278
Closes-Bug: #1611066