Updated openstack/openstack Project: openstack/ironic  ed0b7c672fd14b61a94ca6e80aa4e3641e0e1705
Officially deprecate agent passthru classes and API

We've introduced the new set of API endpoints for lookup and heartbeat
which all drivers should implement now and all ramdisk should use.

IPA switched to the new API in the Newton release. The deprecated passthru
endpoints and classes will be removed in the Ocata release.

Change-Id: If0d189ecdd4b84d1dbac274e6dec200e13dfe37c
Depends-On: I7160c99ca63585fc333482fa578fdf5e0962b2b6
Closes-Bug: #1570841