Updated openstack/openstack Project: openstack/python-novaclient  35057e1a7379bd495f97013978ae9334143f77bd
Fix boot --nic error message for v2.32 boundary

If the --nic args were incorrect the error message said
that nic tags were supported but that's dependent on the
microversion used. This fixes the error message to be
conditional based on the 2.32 microversion. It also pulls
the error message generation out of the loop since it's
static.

Change-Id: I5d71ba7c490aa96f72094398b80af85fcde21ac6
Closes-Bug: #1611454