Updated openstack/openstack Project: openstack-infra/devstack-gate  dcd690265146e89c363acd1b070644c4eb1f82fe
Add sanity check on devstack generated tempest.conf

There are some parts of devstack that call the service api to verify
the extension list we pass to tempest. This is fine for the out of
gate case if it makes life easier for users. But, we should not ever
trust this for gate testing. This commit adds a check on master to
ensure we always are using the default value of 'all'. On stable we
hard code a list of available extensions so this isn't a concern
there.

Change-Id: I01be3c1392fd4642bdd4633ba5624ddacb608bc5