Updated openstack/openstack Project: openstack/neutron  0e1e93f5c4fcd5b3042d4fb8691e3c30e745d819
Relax bound constraint for trunk parent ports

This patch revisits the is_bound() condition for ports that
are about to be used as parent in a trunk. The existing
logic checked that for a port to be bound it is either an
unbound compute port or it is associated to a host. In pratice
only the latter suffices.

Partially-implements: blueprint vlan-aware-vms

Change-Id: I1c2c877b5cc421e05d97c7f528546a6a2c14aa46