Updated openstack/openstack Project: openstack/neutron  8da4d0d471aa73f4563d64aa33dac85bea943477
Delete HA network if last HA router is migrated

If an HA router is deleted and it's the last one a tenant has, the HA
network is deleted as it's no longer needed. The same should occur when
migrating the last HA router of a tenant from HA to Legacy.

Closes-Bug: #1610645
Change-Id: Ib27224c028bef98b78a5cea91e62ea98a2847008