Updated openstack/openstack Project: openstack/glance-specs  a6cfff857ded915787c359c3282fdbc1e166b9a8
Start Ocata specs proposals

Some of the Newton specs did not get enough attention from cores
during the cycle. Let's start the Ocata folder for starting planning on
the specs for next cycle.

This needs to be started/merged in Newton for people to start planning
on writing specs and then help the glance team and community
determine/plan work for Ocata.

Co-Authored-By: Nikhil Komawar <nik.komawar@gmail.com>
Co-Authored-By: Brian Rosmaita <brian.rosmaita@rackspace.com>

Change-Id: I16d1dd18c3efceb2245239f9c95483e717c70c4d