Updated openstack/openstack Project: openstack/nova  b01f359ffd40014be10d0d54563fae31a1b3119a
virt/hardware: Check for threads when "required"

The 'require' case "requires" the presence of hardware threads on a
host. At present, this check is done using the NUMATopology filter.
Unfortunately, this means that if this filter is disabled then
instances can be scheduled on invalid hosts. Resolve this by adding a
new check to be run when hosts are actually scheduling.

Change-Id: Ia9e4784e02ca9ce7a3d81c962b95bee100f6db42
Closes-bug: #1550269