Updated openstack/openstack Project: openstack-infra/project-config  e868f1845384ab61af30aeff020f8c281196f24a
try harder to update constraint when a library is released

Some of the libraries are listed in the constraints file using the name
found in setup.cfg and some are listed using their canonical name. There
doesn't seem to be any pattern to which is which, so just try to update
both names under the assumption that only one will be present or if for
some unforeseen reason both are there that we want them to be the same
anyway.

Change-Id: Ia02a4c4365f1e0eb60f186c12a0c16849030c8a5
Signed-off-by: Doug Hellmann <doug@doughellmann.com>