Updated openstack/openstack Project: openstack-infra/project-config  b448daa6a02bf23a8203e8c78815a97974d0a3ad
Remove the ironic-parallel-nv from tempest experimental

The ironic-parallel-nv job on tempest does not actually run any tests.
My guess is that it's designed to be used with ironic's tempest plugin
(based on the selection regex used to cause [1]) But, since tempest
doesn't include that plugin the job doesn't have anything to run and
fails.

[1] http://logs.openstack.org/56/347756/11/experimental/gate-tempest-dsvm-ironic-parallel-nv/3aa0f97/console.html#_2016-08-09_21_49_19_734536

Change-Id: Ie0b952217e20510aa538f6dec50a60c9361f4751