Updated openstack/openstack Project: openstack-infra/subunit2sql  3ea602d94d02e25542f72c280d0f6428a25c1933
Add DB API function to get list of runs from ids

This commit adds a new function to the DB api for getting a list of
multiple runs from an input list of run_ids. This avoids the need to
loop over a list and call get_run_by_id().

Change-Id: I5df793ec5417245101a6e016def3bddfe7fb25e7