Updated openstack/openstack Project: openstack-infra/storyboard-webclient  5b17e47f0cc58ebc1590a2566e827a21d6536820
Comment some settings in worklist resource

There were a couple of mysterious 'true, false' settings, and it was
not clear what they toggled when someone looked at this file alone. The
answer lay in resourcefactory.js, which is in another directory
(src/app/services/service). This patch documents them in the same
file that they're used.

Change-Id: Ia645c9ddf42656db60cd10ba3351c65ac275ef53