Updated openstack/openstack Project: openstack/nova  1a6df701974f177674d856dca4a97dc9d52432fc
rt: Refactor test_dupe_filter unit test

The TestMoveClaim.test_dupe_filter() unit test was a poorly-written unit
test. Not only was it not actually isolating the particular piece of
code being tested properly, but the assertion it was making was not
proving that filtering that the code was doing actually worked.

This patch refactors that unit test out of TestMoveClaim and into its
own TestUpdateUsageFromMigrations test case class that isolates the area
of tested code to just the function in question and puts assertions into
the unit test that actually verify the *correct* migration is filtered
not just *any* migration.

Change-Id: Icff202e0c8866cad72fe977281a5829759659282