Updated openstack/openstack Project: openstack/nova  30a9e69bcdc9fa096474f195efcb2b7a1b59944e
rt: Remove incorrect unit test of resize revert

There was one remaining "unit" test in test_tracker.py's TestMoveClaim
class: test_revert_reserve_source(). This test wasn't a unit test -- it
was emulating interactions across both a source and destination resource
tracker, blurring the unit-tested code surface area. But more
importantly, this test had a serious flaw in it. In order to "emulate"
the resize operation on the destination host, the test was calling
instance_claim(). However, for resizes, the RT's resize_claim() is the
method that is used on the destination host. Therefore, the assertions
immediately after the use of instance_claim() on the destination
resource tracker were entirely meaningless.

There were already lots of existing unit tests in test_tracker's
TestUpdateAvailableResource that validated the resource usage accounting
on hosts that had incoming, outgoing, and same-host migrations so there
was no point keeping this non-unit unit test around.

Change-Id: I219f3de0240759c0b8b367828365c4df8bb4986d