Updated openstack/openstack Project: openstack/openstack-manuals  9bd7ed25c0663d4ddafcdb3dce734132f9ef2048
www: Use Redirect directive instead of RedirectMatch for exact links

If regular expression match for URL is unnecessary,
there is no reason to use RedirectMatch.
We can also avoid unnecesary/unexpected redirection
and it is better from performance perspective.

This commit also updates the existing redirct related to
the admin guide so that we can avoid double redirection.

Change-Id: Ib5df7a9255e26fad3364a08daf7cec0f23c8aa65