Updated openstack/openstack Project: openstack/cinder  d2f612c91b4180ad818d3c5d0ca33f8f3d1948f3
Managed snapshot should return ceiling of size in Huawei driver

The function manage_existing_snapshot_get_size returns the floor of
volume snapshot size in huawei driver. But in the driver API:"When
calculating the size, round up to the next GB." This patch fixes it
by returning the ceiling of volume snapshot size.

Change-Id: I75dac77ef6d13a7a217569f5b060a198461d344f
Closes-Bug: #1591597