Updated openstack/openstack Project: openstack-infra/project-config  5dd67b95593b286809b09077cca9a6c50c020d97
Increase rh1 max-servers to 75

The rh1 cloud was initially brought up with only 50 servers.  Now
that we've had a chance to work out some bugs and it seems to be
running stable, we can increase the available servers.  For now,
a 50% increase to 75 seems like it should be safe and still well
within our capacity.

Change-Id: I8e9bac8c48c068cbd513190dfcebaf0e297f2dba