Updated openstack/openstack Project: openstack/cinder  be11e183d8ef97a4c3003c637800e383dea9929a
Prevent doc generation failure on OVO decorators

Currently Sphinx will fail to generate the documentation if we are using
decorators from any OVO in cinder.objects, because the OVOs are only
added to the cinder.objects namespace when the CLI programs are run.

So we need to run `register_all` method during documentation generation
to avoid failures like:

    AttributeError: 'module' object has no attribute 'Volume'

This patch modifies autogenerated doc/source/conf.py file and since now
it is no longer completely autogenerated it is added to PEP8 checks,
which required some minor changes.

Change-Id: Ifeeef61a778f9e3f3daceba8ed05cd2036219499