Updated openstack/openstack Project: openstack/cinder  5d6c51ea519fed71aeebbfd07b0dad9f2f9a987d
Storwize: Quote vdisk names to handle spaces

For Storwise SSH commands that take one or more vdisk names,
this fix quotes the vdisk parameter name in the common helper
methods. Passing a name with a space in it causess an
SSHThreatInjection exception to be raised.

This patch avoids the exception
by quoting the command line parameter in the same
fashion as was used to quote the host parameter. The UT module
was also updated to strip the quotes such that the volumes oculd
be looked up by the simulator. This mimics what SSH shell would be
doing in stripping quotes.

Change-Id: I329b22ba9006b36871ceb206aadbe51502947f54
Closes-bug: #1602008