Updated openstack/openstack Project: openstack-infra/project-config  ab89ab40ed74db306ce10e36341d39f23231f012
Properly set up UPPER_CONSTRAINTS_FILE for translations

For the translation scripts, UPPER_CONSTRAINTS_FILE was not set anywhere, so
rework the logic and set it up ourselves - like run-tox.sh and other
scripts do.

Also, rename UPPER_CONSTRAINTS to UPPER_CONSTRAINTS_FILE as that is the
proper name for these.

Change-Id: If3f2eab1ee04bd7d53b714f5abdb1a3a2f1be266