Updated openstack/openstack Project: openstack/cinder  2ebd8b1275515a68ddc29030243392d8aa326584
CI: Add CI_WIKI_NAME to all drivers

This patch adds a CI_WIKI_NAME to each driver object.  The value is the exact
name of the ThirdPartySystems wiki page.   This allows us to create an
automated tool to associated jobs to drivers and track their CI reporting
status correctly.

This patch also updates the generate_driver_list.py script to output the
driver list as a python list of dicts that can be directly consumed.

Change-Id: I0ec5f705e91f680a731648cf50738ea219565f70