Updated openstack/openstack Project: openstack/neutron  4e276c3b477727680bb86d5e1f0e588cc541b696
Change external_network_bridge default to ''

This changes the default for external_network_bridge to '' to
encourage operators to deploy L3 agents in the correct manner.

This patch also adds a functional test to ensure that namespaces
with an existing external gateway interface are not torn down and
rewired on change.

Closes-Bug: #1563070
Change-Id: If533cf7c4c379be78f5a15073accaff7f65973ab