Updated openstack/openstack Project: openstack/neutron  6c7dd804d32211a1eb39cd336217c04ab2a74047
lb-agent: handle exception when bridge slave already removed

An exception can happen when a network is deleted because the
lb-agent tries to removes the dhcp tap from the bridge at about
the same time as the dhcp-agent is deleting the tap. The unhandled
exception means the bridge does not get deleted and a log error.

Closes-Bug: #1611612
Change-Id: Ia9a6b5fc49e239769e850e9486454e81e3a4b96f