Updated openstack/openstack Project: openstack/neutron  2a05b28c05c6509db9b5b3e543df33373f2f996b
Fix duplicate routerport handling

Change 6281fdd introduced a try: except to catch
a DuplicateError introduced by it's new constraint.
However, the try moved it outside of the context manager
so it broke the port cleanup logic.

This patch completely eliminates the catch for the
duplicate entry since we retry those anyway which
would let the regular check see the duplication.

This also adds a test to prevent another regression of
being moved outside of the context manager.

Related-Bug: #1535551
Related-Bug: #1600344

Change-Id: I5f473fff4f8372852d563c79dac2991089eb0b77