Updated openstack/openstack Project: openstack/nova  04bc62e4a061000962dce34d54d4e65e239ffa01
vnc host options need to support hostnames

When updating the config options the VNC options were switched from
StrOpt to IPOpt. However these are hostnames, they even say so in the
option name, so IPOpt is too restrictive, and could break folks in
upgrade if they set these to hostnames.

Change-Id: Ib2062407dcf9cba8676b0f38aa0c63df25cc7b38
Closes-Bug: #1611940