Updated openstack/openstack Project: openstack-infra/project-config  fd31df6739f1be4eec273a9154189bd8a97ecaca
Add jobs for Manila glusterfs, glusterfs-native drivers with Heketi layout

"Heketi layout" is a novel setup method for the Manila gluster* drivers
which require dedicated Jenkins jobs.

Set the job as experimental and non-voting for the Manila project
as the Heketi layout is still WIP.

Set the job as non-voting for the devstack-plugin-glusterfs project until
the job is proven to be stable.

Change-Id: Ic3a358f5f8a0f5ea87989c87d8e246c426b937c2