Updated openstack/openstack Project: openstack-infra/shade  19d36dd768b57e1f0632e77f07480a8e40914043
Add ability to upload duplicate images

In the basic case, avoiding re-uploading images is nice because
uploading images is a costly operation. However, there are also usecases
for uploading new images with the same name as an operator which are
quite valid. Support these by allowing for a flag that overrides the
duplicate checking logic.

Change-Id: I6a00753654d500d68750e46c5ec2423a36903552