Updated openstack/openstack Project: openstack/tripleo-heat-templates  44b67ebdeb179af66dad78c2c54eb07723f46d4d
Reinstate CloudDomain properties for Controller/Compute roles

These were removed in https://review.openstack.org/#/c/347050
but it turns out the defaults in the role templates is bad, as
an empty string results in a malformed hosts file fqdn.

So, partially revert that patch so we always pass the global
CloudDomain from overcloud.yaml, accepting the default configured
there, and remove the empty-string defaults in the role templates.

Change-Id: I0ea4190a23488986a3ee9e887328e0e7a03fe3aa