Updated openstack/openstack Project: openstack/cinder  5f65ed75ba59f87c4d75289568dbcb30902aee1e
Add strict Boolean checking for volume create

There is no strict boolean checking for the parameter
"multiattach" of API /volumes, so that
any invalid boolean value can be specified.
This patch adds a strict checking for it to prevent
invalid value, and adds tests for this change as well.

Change-Id: I2a5dc45a3b238a2d875d7fbf7159eb91d1f111bf
Partial-Bug: #1594261