Updated openstack/openstack Project: openstack/nova  754f18eb5ec309f1cf71fe5ce252f4ab98936ade
config options: improve help text of database (related) options (1/2)

This change enhances the help texts of the database related config
options. They are in the "DEFAULT" section of the "nova.conf" and
get used on the DB layer but describe a usage on the "compute" level.
They should probably live in another section than "DEFAULT".

bp centralize-config-options-newton

Change-Id: I5caeff8bcb5f0b63f5ec4e79ab5f6cde16ae018f