Updated openstack/openstack Project: openstack/nova  602006a80d27f39eb8ba81b30fb0cf8b64eb9e57
Fix handling of status in placement API json_error_formatter

The API-WG standard for error responses[1] says that the value of the
'status' key should be an int. webob error formatters receive the
status as a string (for example '400 Bad Request'). This change
fixes that and adds unit test coverage for the json_error_formatter
method.

This was a problem that was highlighted during review of the basic
framing of the placement API but was not resolved before that code
merged.

[1] http://specs.openstack.org/openstack/api-wg/guidelines/errors.html

Change-Id: I243eeaf62f83417b77bc6fa4a7d6f898b1382ac1
Partially-Implements: blueprint generic-resource-pools