Updated openstack/openstack Project: openstack/glance  ef41425217ff04be5f50fe9796df98e5435bf213
Add test class to versions tests

The current versions tests contain a class to test the versions
response, and a class to test version negotiation, but this put us
in a situation where the 'versions' response was correctly updated
for v2.3, but the negotiation to map v2.3 -> v2 was not, thereby
making calls to /v2.3/whatever fail even though all tests in both
classes passed.

This patch adds a new test class that inherits from both of the
current test classes.  It contains tests to verify that all the
versions mentioned in the versions response are mapped correctly
so this kind of partial update situation won't happen again.

Partial-Bug: #1609571
Change-Id: I9e93b0f075eb6129c912dfca0a5d005f71bbc253
Co-Authored-By: Nikhil Komawar <nik.komawar@gmail.com>
Co-Authored-By: Brian Rosmaita <brian.rosmaita@rackspace.com>